requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
numpy==1.19.4
pandas==1.1.5
pkg-resources==0.0.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.5
rsa==4.6
six==1.15.0
Telethon==1.18.2
tornado==6.1
tzlocal==2.1
